create define_package_managers macro
